+++ test/oauth2_flows/implicit/gapi_auth_immediate_test.dart
@@ -10,7 +10,7 @@ import 'package:googleapis_auth/src/utils.dart' as utils;
 import 'utils.dart';
 
 main() {
-  impl.GapiUrl = resource('gapi_auth_force.js');
+  impl.GapiUrl = resource('gapi_auth_immediate.js');
 
   test('gapi-auth-force', () {
     var clientId = new auth.ClientId('foo_client', 'foo_secret');
@@ -18,7 +18,7 @@ main() {
 
     auth.createImplicitBrowserFlow(clientId, scopes)
         .then(expectAsync((auth.BrowserOAuth2Flow flow) {
-      flow.obtainAccessCredentialsViaUserConsent(forceUserConsent: true)
+      flow.obtainAccessCredentialsViaUserConsent(immediate: true)
           .then(expectAsync((auth.AccessCredentials credentials) {
         var date = new DateTime.now().toUtc().add(const Duration(
             seconds: 3210 - utils.MAX_EXPECTED_TIMEDIFF_IN_SECONDS));