+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0
+- Rename forceUserConsent parameter to immediate.
+- Add runHybridFlow() function to auth_browser, plus corresponding
+  HybridFlowResult class
+
 ## 0.1.1
 - Add clientViaApiKey() functions to auth_io/auth_browser.